requirements.txt
# Core
pytz==2020.1
Django==3.1
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.16.0.145

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.11.1
Markdown==3.2.2
django-filter==2.3.0

# Developer Tools
ipdb==0.13.3
ipython==7.17.0
mkdocs==1.1.2
flake8==3.8.3
django-extensions==3.0.5

# Testing
mock==4.0.2
factory-boy==2.12.0
django-nose==1.4.6
nose-progressive==1.5.2
coverage==5.2.1

# Static and Media Storage
django-storages==1.9.1
boto3==1.14.40

